# Remove yaml require for 5% init speedup