@apainintheneck 2023-04-18
I ran into this a few weeks ago but I forgot to open a PR for it. I believe `stringio` isn't used in this file either and can probably just be removed while you're here.