@carlocab 2023-04-18
I'm more surprised there isn't an `include-what-you-use` type linter that can tell you about useless `require`s. But I guess that's a much harder problem for interpreted languages.