@dduugg 2023-04-18
> I'm more surprised there isn't an `include-what-you-use` type linter that can tell you about useless `require`s. But I guess that's a much harder problem for interpreted languages.

Yeah, the unfortunate reality is that `require`s often include monkey patches or other side effects that make this challenging. I think that's essentially what @apainintheneck ran into when [researching](https://github.com/Homebrew/brew/pull/15166) using `autoload` as well.)